+++ packages/happner-2/RELEASES.md
@@ -1046,4 +1046,9 @@ SMC-1645: made error messages more consistent, encapsulating mesh.component.meth
 -----------------
   - feat: happn-3 changePassword in client
   - feat: happn-3 resetPassword supported by auth provider
-  
+
+13.0.0 2023-05-20
+-----------------
+  - feat: Happner-2 security change password
+  - feat: Happner-2 security reset password as anonymouse
+  - feat: deprecate updateOwnUser (breaking)